pyproject.toml
[project]
name = "feedhound"
version = "0.1.0"
description = ""
authors = [
    {name = "NumiViX",email = "yakhin.v.r@yandex.ru"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[all] (>=0.115.12,<0.116.0)",
    "sqlalchemy[asyncio] (>=2.0.40,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "pydantic (>=2.11.4,<3.0.0)",
    "aiohttp (>=3.11.18,<4.0.0)",
    "feedparser (>=6.0.11,<7.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "redis (>=4.5.4,<5.0.0)",
    "celery[redis] (>=5.5.2,<6.0.0)",
    "alembic (>=1.15.2,<2.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "pydantic-settings (>=2.9.1,<3.0.0)",
    "python-jose[cryptography] (>=3.4.0,<4.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
